move interfaces to their own module
