+++ stanza/nlp/__init__.py
@@ -1,0 +1,2 @@
+from data import *
+from corenlp import *
